agentbuy/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agentbuy"
version = "0.1.0"
description = "Universal AI Purchase Agent - Voice/text-first autonomous purchasing"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "Proprietary"}
authors = [
    { name = "AgentAuth Team" }
]
keywords = ["ai", "agent", "purchasing", "voice", "automation"]

dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.5.0",
    "httpx>=0.26.0",
    "openai>=1.10.0",
    "python-multipart>=0.0.6",
    "agentauth-client>=0.1.0",
    "playwright>=1.41.0",
    "python-dotenv>=1.0.0",
    "websockets>=12.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "black>=24.1.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
]

[project.scripts]
agentbuy = "src.api.main:run"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "N", "W"]

[tool.mypy]
python_version = "3.11"
strict = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
